fixed imports and list comperhension syntax
