+++ integration_testing/run_tests.py
@@ -288,6 +288,6 @@ def test_multilabel_prediction():
 
 
 setup_testing_logger()
-test_timeseries()
-#test_multilabel_prediction()
+#test_timeseries()
+test_multilabel_prediction()
 #test_one_label_prediction()
