+++ mindsdb/libs/model_examination/column_evaluator.py
@@ -1,4 +1,3 @@
-import pickle
 
 
 class ColumnEvaluator():
@@ -11,21 +10,6 @@ class ColumnEvaluator():
     def __init__(self):
         self.columnless_predictions = {}
         self.normal_predictions = None
-
-    def pickle(self):
-        """
-        Returns a version of self that can be serialized into mongodb or tinydb
-        :return: The data of a HypothesisExecutor serialized via pickle and decoded as a latin1 string
-        """
-        return pickle.dumps(self).decode(encoding='latin1')
-
-    @staticmethod
-    def unpickle(pickle_string):
-        """
-        :param pickle_string: A latin1 encoded python str containing the pickle data
-        :return: Returns a HypothesisExecutor object generated from the pickle string
-        """
-        return pickle.loads(pickle_string.encode(encoding='latin1'))
 
     # @TODO: Move into a more generic file
     @staticmethod
@@ -49,7 +33,7 @@ class ColumnEvaluator():
 
         for input_column in input_columns:
             # See what happens with the accuracy of the outputs if only this column is present
-            ignore_columns = [col if col != input_column for col in input_columns]
+            ignore_columns = [col for col in input_columns if col != input_column ]
             col_only_predictions = model.predict('validate', ignore_columns)
             col_only_accuracy = evaluate_accuracy(self.normal_predictions, full_dataset, stats, input_columns)
             col_only_normalized_accuracy = col_only_accuracy/normal_accuracy
